requirements.txt
python-musicpd
wxPython